SAM-adaptor/code/requirements.txt
imageio==2.9.0
ipython==7.34.0  
matplotlib>=3.4  
opencv-python~=4.7.0.72
PyYAML~=6.0
scikit-learn~=1.2.2
scipy~=1.10.1
tqdm~=4.51.0
numpy>=1.21.2
typing~=3.7.4.3
terminaltables~=3.1.10
Pillow~=9.4.0
tensorboardX~=2.6
onnxruntime~=1.14.1
setuptools~=67.6.1
timm~=0.3.2
easydict~=1.10
attr~=0.3.2
thop~=0.1.1.post2209072238
torchsummary~=1.5.1
